First attempt at Python bindings
