+++ tests/CMakeLists.txt
@@ -54,7 +54,17 @@ add_test(
     pytaylor_import
   COMMAND
    ${CMAKE_COMMAND} -E env PYTAYLOR_MODULE_PATH=${STAGED_INSTALL_PREFIX}/${INSTALL_LIBDIR}
-   ${PYTHON_EXECUTABLE} -m pytest -rws -v --capture=sys ${CMAKE_CURRENT_SOURCE_DIR}/test_taylor_import.py
+   ${PYTHON_EXECUTABLE} -m pytest -rws -v --capture=sys ${CMAKE_CURRENT_SOURCE_DIR}/test_import.py
+  WORKING_DIRECTORY
+    ${CMAKE_CURRENT_BINARY_DIR}
+  )
+
+add_test(
+  NAME
+    pytaylor_univariate
+  COMMAND
+   ${CMAKE_COMMAND} -E env PYTAYLOR_MODULE_PATH=${STAGED_INSTALL_PREFIX}/${INSTALL_LIBDIR}
+   ${PYTHON_EXECUTABLE} -m pytest -rws -v --capture=sys ${CMAKE_CURRENT_SOURCE_DIR}/test_univariate.py
   WORKING_DIRECTORY
     ${CMAKE_CURRENT_BINARY_DIR}
   )
