--- tests/test.py
@@ -1,3 +1,0 @@
-from pytaylor import taylor
-
-taylor1 = taylor(1.0)
